Versione finale refactoring dbandbusiness Gruppo 6
