+++ DataLayer/ClassesManage.cs
@@ -12,11 +12,9 @@ namespace SchoolGrades.DataLayer
 {
     class ClassesManage
     {
-        //Cesare Colella, Francesco Citarella, Andrea Siboni, Riccardo Brunelli 4L
+        //Andrea Siboni, Francesco Citarella, Cesare Colella, Riccardo Brunelli 4L
 
         DataLayer dl = new DataLayer();
-
-
 
         internal void UpdatePathStartLinkOfClass(Class currentClass, string text)
         {
@@ -440,7 +438,237 @@ namespace SchoolGrades.DataLayer
         }
 
 
+        internal void EraseClassFromClasses(Class Class)
+        {
+            //EraseAllStudentsOfAClass(Class); 
+            using (DbConnection conn = dl.Connect())
+            {
+                // delete all the references in link table between students and classes
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Classes_Students" +
+                    " WHERE Classes_Students.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // erase class from Classes_SchoolSubjects
+                cmd.CommandText = "DELETE FROM Classes_SchoolSubjects" +
+                    " WHERE Classes_SchoolSubjects.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // erase class from Classes_Tests
+                cmd.CommandText = "DELETE FROM Classes_Tests" +
+                    " WHERE Classes_Tests.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // erase class from table Classes 
+                cmd.CommandText = "DELETE FROM Classes" +
+                    " WHERE Classes.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
 
+        internal string CreateOneClassOnlyDatabase(Class Class)
+        {
+            string newDatabasePathName = Class.PathRestrictedApplication + "\\SchoolGrades\\Data\\";
+            if (!Directory.Exists(newDatabasePathName))
+                Directory.CreateDirectory(newDatabasePathName);
+
+            string newDatabaseFullName = newDatabasePathName +
+                System.DateTime.Now.ToString("yyyy-MM-dd_HH.mm.ss") +
+                "_" + Class.Abbreviation + "_" + Class.SchoolYear + "_" +
+                Commons.FileDatabase;
+            File.Copy(Commons.PathAndFileDatabase, newDatabaseFullName);
+
+            // open a local connection to database 
+            DataLayer newDatabaseDl = new DataLayer(newDatabaseFullName);
+
+            // erase all the data of the students of other classes
+            using (DbConnection conn = newDatabaseDl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+
+                // erase all the other classes
+                cmd.CommandText = "DELETE FROM Classes" +
+                " WHERE idClass<>" + Class.IdClass + ";";
+                cmd.ExecuteNonQuery();
+
+                // erase all the lessons of other classes
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Lessons" +
+                    " WHERE idClass<>" + Class.IdClass + ";";
+                cmd.ExecuteNonQuery();
+
+                // erase all the students of other classes from the link table
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Classes_Students" +
+                 " WHERE idClass<>" + Class.IdClass + ";";
+                cmd.ExecuteNonQuery();
+
+                // erase all the students of other classes 
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Students" +
+                    " WHERE idStudent NOT IN" +
+                    " (SELECT idStudent FROM Classes_Students);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the StartLinks of other classes
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Classes_StartLinks" +
+                    " WHERE idClass<>" + Class.IdClass + ";";
+                cmd.ExecuteNonQuery();
+
+                // erase all the grades of other classes' students
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Grades" +
+                    " WHERE idStudent NOT IN" +
+                    " (SELECT idStudent FROM Classes_Students);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the links to photos of other classes' students
+                // !! retains previous year's photos of this classes students !!
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM StudentsPhotos_Students" +
+                    " WHERE idStudent NOT IN" +
+                    " (SELECT idStudent FROM Classes_Students);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the annotations of other classes
+                cmd.CommandText = "DELETE FROM StudentsAnnotations" +
+                    " WHERE idStudent NOT IN" +
+                    " (SELECT idStudent FROM Classes_Students)" +
+                    ";";
+                cmd.ExecuteNonQuery();
+
+                // erase all the photos of other classes' students
+                // !! retains previous year's photos of this classes students !!
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM StudentsPhotos" +
+                    " WHERE idStudentsPhoto NOT IN" +
+                    " (SELECT idStudentsPhoto FROM StudentsPhotos_Students);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the questions of the students of the other classes
+                // !! StudentsQuestions currently not used !!
+                cmd.CommandText = "DELETE FROM StudentsQuestions" +
+                    " WHERE idStudent NOT IN" +
+                    " (SELECT idStudent FROM Classes_Students);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the answers  of the students of the other classes
+                // !! StudentsAnswers currently not used !!
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM StudentsAnswers" +
+                " WHERE idStudent NOT IN" +
+                " (SELECT idStudent FROM Classes_Students);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the tests of students of the other classes
+                // !! StudentsTests currently not used !!
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM StudentsTests" +
+                " WHERE idStudent NOT IN" +
+                " (SELECT idStudent FROM Classes_Students);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the images of other classes' lessons
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Lessons_Images" +
+                    " WHERE idLesson NOT IN" +
+                    " (SELECT idLesson from Lessons);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the topics of other classes' lessons
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Lessons_Topics" +
+                    " WHERE idLesson NOT IN" +
+                    " (SELECT idLesson from Lessons);";
+                cmd.ExecuteNonQuery();
+
+                // erase all the users
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Users" +
+                    ";";
+                cmd.ExecuteNonQuery();
+
+                // copy all the students' photo files that aren't already there or that have a newer date 
+                string query = "SELECT StudentsPhotos.photoPath" +
+                " FROM StudentsPhotos" +
+                " JOIN StudentsPhotos_Students ON StudentsPhotos_Students.idStudentsPhoto = StudentsPhotos.idStudentsPhoto" +
+                " JOIN Classes_Students ON StudentsPhotos_Students.idStudent = Classes_Students.idStudent" +
+                " WHERE Classes_Students.idClass = " + Class.IdClass + "; ";
+                cmd = new SQLiteCommand(query);
+                cmd.Connection = conn;
+                DbDataReader dReader = cmd.ExecuteReader();
+                while (dReader.Read())
+                {
+                    string destinationFile = Class.PathRestrictedApplication + "\\SchoolGrades\\Images\\" + (string)dReader["photoPath"];
+                    if (!Directory.Exists(Path.GetDirectoryName(destinationFile)))
+                    {
+                        Directory.CreateDirectory(Path.GetDirectoryName(destinationFile));
+                    }
+                    if (!File.Exists(destinationFile) ||
+                        File.GetLastWriteTime(destinationFile)
+                        < File.GetLastWriteTime(Commons.PathImages + "\\" + (string)dReader["photoPath"]))
+                        try
+                        {
+                            // destination file not existing or older
+                            File.Copy(Commons.PathImages + "\\" + (string)dReader["photoPath"],
+                                destinationFile);
+                        }
+                        catch { }
+                }
+                // copy all the picture's files that aren't already there or that have a newer date 
+                query = "SELECT Images.imagePath, Classes.pathRestrictedApplication" +
+                    " FROM Images" +
+                    " JOIN Lessons_Images ON Lessons_Images.idImage=Images.idImage" +
+                    " JOIN Lessons ON Lessons_Images.idLesson=Lessons.idLesson" +
+                    " JOIN Classes ON Classes.idClass=Lessons.idClass" +
+                    " WHERE Lessons.idClass=" + Class.IdClass +
+                    ";";
+                cmd = new SQLiteCommand(query);
+                cmd.Connection = conn;
+                dReader = cmd.ExecuteReader();
+                while (dReader.Read())
+                {
+                    if (dReader["pathRestrictedApplication"] is DBNull)
+                    {
+                        Console.Beep();
+                        break;
+                    }
+                    if (dReader["imagePath"] is DBNull)
+                    {
+                        Console.Beep();
+                        break;
+                    }
+                    string destinationFile = (string)dReader["pathRestrictedApplication"] +
+                        "\\SchoolGrades\\" + "Images" + "\\" + (string)dReader["imagePath"];
+                    if (!Directory.Exists(Path.GetDirectoryName(destinationFile)))
+                    {
+                        Directory.CreateDirectory(Path.GetDirectoryName(destinationFile));
+                    }
+                    if (!File.Exists(destinationFile) ||
+                        File.GetLastWriteTime(destinationFile)
+                        < File.GetLastWriteTime(Commons.PathImages + "\\" + (string)dReader["imagePath"]))
+                        // destination file not existing or older
+                        try
+                        {
+                            File.Copy(Commons.PathImages + "\\" + (string)dReader["imagePath"],
+                                destinationFile);
+                        }
+                        catch { }
+                }
+                dReader.Dispose();
+                // compact the database 
+                cmd.CommandText = "VACUUM;";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+            // ???? Students_GradeTypes serve ???? se non serve, eliminare
+            // ???? StudentsTestsStudentsTests serve ???? se non serve, eliminare
+            return Class.PathRestrictedApplication;
+        }
