+++ DataLayer/GradeData.cs
@@ -1,0 +1,73 @@
+﻿using System;
+using System.Data.Common;
+using System.Data;
+using System.Data.SQLite;
+using System.IO;
+using System.Collections.Generic;
+using gamon;
+using System.Windows.Forms;
+using SchoolGrades.DbClasses;
+
+namespace SchoolGrades.DataLayer
+{
+    //Andrea Siboni, Francesco Citarella, Cesare Colella, Riccardo Brunelli 4L
+    class GradeData
+    {
+        DataLayer dl = new DataLayer();
+
+        internal GradeType GetGradeType(string IdGradeType)
+        {
+            GradeType gt = null;
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * FROM GradeTypes";
+                cmd.CommandText += " WHERE idGradeType ='" + IdGradeType + "'";
+                cmd.CommandText += ";";
+                dRead = cmd.ExecuteReader();
+                dRead.Read();
+                gt = GetGradeTypeFromRow(dRead);
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return gt;
+        }
+        private GradeType GetGradeTypeFromRow(DbDataReader Row)
+        {
+            if (Row.HasRows)
+            {
+                GradeType gt = new GradeType();
+                gt.IdGradeType = (string)Row["idGradeType"];
+                gt.IdGradeTypeParent = SafeDb.SafeString(Row["IdGradeTypeParent"]);
+                gt.IdGradeCategory = (string)Row["IdGradeCategory"];
+                gt.Name = (string)Row["Name"];
+                gt.DefaultWeight = (double)Row["DefaultWeight"];
+                gt.Desc = (string)Row["Desc"];
+                return gt;
+            }
+            return null;
+        }
+
+        internal List<GradeType> GetListGradeTypes()
+        {
+            List<GradeType> lg = new List<GradeType>();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * FROM GradeTypes;";
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    GradeType gt = GetGradeTypeFromRow(dRead);
+                    lg.Add(gt);
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+                return lg;
+            }
+        }
+
+    }
+}
